requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
cbor2==5.8.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
einops==0.8.1
fastapi==0.128.0
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httptools==0.7.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
iopath==0.1.10
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modal==1.3.0.post1
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.4
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMatting==1.1.14
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
spandrel==0.4.1
referencing==0.37.0
regex==2026.1.15
rembg==2.0.69
requests==2.32.5
resend==2.19.0
rich==14.2.0
rpds-py==0.30.0
safetensors==0.7.0
sam2==1.1.0
scikit-image==0.26.0
scipy==1.17.0
scour==0.38.2
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
sympy==1.14.0
synchronicity==0.11.1
tifffile==2026.1.14
timm==1.0.24
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.6
typer==0.21.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
vtracer==0.6.11
watchfiles==1.1.1
websockets==16.0
yarl==1.22.0
